Ontroduce install command
